This commit introduces a Pratt parser for handling shell commands, replacing the previous simple parsing approach. The changes include:

- Implementation of Pratt parsing algorithm for operator precedence and associativity.
- Creation of Abstract Syntax Tree (AST) to represent the parsed command structure.
- Tokenization of input strings into tokens with associated types and binding powers.
- Functions for handling different token types (words, operators, quotes, redirections).
- AST traversal and command list generation.
- Addition of `ft_isspace` to `libft` and refactoring of `libft` functions to use `size_t` instead of `unsigned int` for indexing and size values.
- Memory management improvements, including `ft_realloc` and dedicated free functions for AST and token structures.
- VSCode configuration files for improved development environment.
- Debugging utilities for printing token arrays, AST structures, and command lists.

+++ libft/ft_putnbrhexmaj.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_putnbrhexmaj.c                                  :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: lengarci <lengarci@student.42.fr>          +#+  +:+       +#+        */
+/*   By: macauchy <macauchy@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/03/16 11:49:00 by lenygarcia        #+#    #+#             */
-/*   Updated: 2025/06/04 10:12:41 by lengarci         ###   ########.fr       */
+/*   Updated: 2025/06/10 11:15:34 by macauchy         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 int	ft_putnbrhexmaj(int fd, int n)
 {
-	unsigned int		count;
-	unsigned int		nb;
+	size_t		count;
+	size_t		nb;
 	char				*hex;
 
 	hex = "0123456789ABCDEF";
