+++ libft/ft_putnbrui.c
@@ -3,18 +3,18 @@
 /*                                                        :::      ::::::::   */
 /*   ft_putnbrui.c                                      :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: lengarci <lengarci@student.42.fr>          +#+  +:+       +#+        */
+/*   By: macauchy <macauchy@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/03/16 11:43:08 by lenygarcia        #+#    #+#             */
-/*   Updated: 2025/06/04 10:11:54 by lengarci         ###   ########.fr       */
+/*   Updated: 2025/06/10 11:15:34 by macauchy         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_putnbrui(int fd, unsigned int n)
+int	ft_putnbrui(int fd, size_t n)
 {
-	unsigned int	count;
+	size_t	count;
 
 	count = 0;
 	if (n >= 10)
